Negative Error
Since this is a class or a type it does not have the ability to print. The most we can really do is return. By returning a Negative Amount Error code the programmer can test for the code and then figure out how they want to deal with it, whether it is printing a message, killing the program or whatever. Essentially, it is not OUR job to really worry about that, all we do is just create the type.