Update Hours
Since it is within the Employee class that we want to update the hours we just want to make the new amount equal the old amount plus the new amount. Remember: The programmers creating the Employee type essentially does NOT KNOW about what the programmers creating the Company type is doing(Even, though you are the one actually making them both :P). Of course, there is collaboration on how the Company wants the Employee type to be made, but they should be fully functional on their own. For example the type str doesn't need to know what the type list does in order to be a str, even though that str will be used in the list