An Employee Class or Company?
So, I removed the records and amount owed from the employee type because that is not the problem of the employee type. The employee is not the one who makes their own record, but the company has a record FOR their employees. The company class would be the one to have the record(which we can have as a dict or list) for their employees. Just like how a bank manages a collection of bank accounts, the company manages a collection of employee records