Objects, Private, and Dictionary
When you are creating a class you have to indicate what it is. In our case, Employee and Company are objects(which is why we have the Company(object)..part). In C++ when we start to go into inheritance this will be where we certain "passed down" traits. 
Quick  example:
class Animal(object):
 ....
...

class Dog(Animal):
....
...

The class animal is an object, which is why in its parenthesis we have object. It is not inheriting any methods. The class Dog is a type of animal, which is why in its parentheses we have Animal. By having animal passed through Dog has access to all of the methods Animal has. Dog can use the animal methods by default(since it is a type of animal) and then in dog we can add new methods or alter the preexisting methods set aside by Animal( but the alteration happen WITHIN dog, NOT animal).

I made the worker records into a dictionary because there will be multiple workers that need to be accounted for. I used a dictionary as opposed to a list because dictionaries are more efficient. Dictionaries are organized as r-b trees(red black trees) so the tree search that would have to do save much more time than a linear search(checking every single one) that would have to be done in a list. Not only that, but dictionaries have key-value pair so we can be specific with EXACTLY which record we want to access.