Hiring a Worker
Since we are using the SSN as the dictionary key-- we are only using the SSN as our key because it is unique-- we just check to see if the SSN is in the dictionary. The reason why we say self before  __workerRec is so that the python interpreter knows what Company object we are referring to( this is good practice because there can be more than one Company object in the same program).